requirements.txt
MetaTrader5 
python-telegram-bot
aioconsole